Rename packages to fit coding standard
