+++ src/main/java/commands/Command.java
@@ -1,9 +1,9 @@
-package UserCommands;
+package commands;
 
 import java.util.Scanner;
 
-import Features.DukeException;
-import Features.TaskList;
+import features.DukeException;
+import features.TaskList;
 
 abstract class Command {
     public abstract TaskList handle(Scanner userScan, TaskList taskList) throws DukeException;
