+++ src/main/java/commands/CommandDeadline.java
@@ -1,12 +1,12 @@
-package UserCommands;
+package commands;
 
 import java.util.Scanner;
 
-import Features.DukeException;
-import Features.TaskList;
-import Features.Ui;
-import Tasks.Deadline;
-import Tasks.Task;
+import features.DukeException;
+import features.TaskList;
+import features.Ui;
+import tasks.Deadline;
+import tasks.Task;
 
 
 /**
