+++ src/test/java/tasks/DeadlineTest.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
