+++ src/test/java/tasks/EventTest.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
